+++ package.json
@@ -25,8 +25,8 @@
     "lint:js": "eslint \"{src,demo}/**/*.{js,jsx,ts,tsx}\"",
     "lint:styles": "stylelint \"{src,demo}/**/*.{css,scss}\"",
     "lint:prettier": "prettier --check \"{src,demo}/**/*.{js,jsx,ts,tsx,css,scss}\"",
+    "check-circular-deps": "node scripts/check-circular-deps.js 48",
     "test": "jest",
-    "check-circular-deps": "node scripts/check-circular-deps.js 71",
     "test:cov": "jest --coverage",
     "test:watch": "jest --watchAll",
     "test:esbuild": "node tests/esbuild-test/esbuild-tester.js",
