+++ src/bundle/MarkdownEditorView.tsx
@@ -16,11 +16,10 @@ import {HorizontalDrag} from './HorizontalDrag';
 import {MarkupEditorView} from './MarkupEditorView';
 import {SplitModeView} from './SplitModeView';
 import {WysiwygEditorView} from './WysiwygEditorView';
-import {MToolbarData, MToolbarItemData} from './config';
 import {useMarkdownEditorContext} from './context';
-import {EditorSettings, EditorSettingsProps} from './settings';
+import {EditorSettings, type EditorSettingsProps} from './settings';
 import {stickyCn} from './sticky';
-import type {WToolbarData, WToolbarItemData} from './toolbar/types';
+import type {MToolbarData, MToolbarItemData, WToolbarData, WToolbarItemData} from './toolbar/types';
 import {getToolbarsConfigs} from './toolbar/utils/toolbarsConfigs';
 import type {MarkdownEditorMode} from './types';
