+++ src/bundle/MarkupEditorView.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import {ClassNameProps, cn} from '../classname';
+import {type ClassNameProps, cn} from '../classname';
 import {ReactRendererComponent} from '../extensions';
 import {logger} from '../logger';
 import {useRenderTime} from '../react-utils/hooks';
@@ -8,8 +8,8 @@ import {useRenderTime} from '../react-utils/hooks';
 import type {EditorInt} from './Editor';
 import {MarkupEditorComponent} from './MarkupEditorComponent';
 import {ToolbarView} from './ToolbarView';
-import type {MToolbarData, MToolbarItemData} from './config/markup';
 import {MarkupToolbarContextProvider} from './toolbar/markup/context';
+import type {MToolbarData, MToolbarItemData} from './toolbar/types';
 import type {MarkdownEditorSplitMode} from './types';
 
 import './MarkupEditorView.scss';
