+++ src/bundle/toolbar/types.ts
@@ -1,12 +1,15 @@
 import type {ActionStorage} from '../../core';
+import type {CodeEditor} from '../../markup';
 import type {ToolbarListButtonData} from '../../toolbar';
 import type {
+    ToolbarButtonPopupData,
     ToolbarData,
     ToolbarGroupData,
     ToolbarGroupItemData,
     ToolbarItemData,
     ToolbarListButtonItemData,
     ToolbarListItemData,
+    ToolbarReactComponentData,
     ToolbarSingleItemData,
 } from '../../toolbar/types';
 
@@ -20,3 +23,12 @@ export type WToolbarGroupItemData = ToolbarGroupItemData<ActionStorage>;
 export type WToolbarListButtonData = ToolbarListButtonData<ActionStorage>;
 export type WToolbarListItemData = ToolbarListItemData<ActionStorage>;
 export type WToolbarListButtonItemData = ToolbarListButtonItemData<ActionStorage>;
+
+export type MToolbarData = ToolbarData<CodeEditor>;
+export type MToolbarItemData = ToolbarItemData<CodeEditor>;
+export type MToolbarSingleItemData = ToolbarSingleItemData<CodeEditor>;
+export type MToolbarGroupData = ToolbarGroupData<CodeEditor>;
+export type MToolbarReactComponentData = ToolbarReactComponentData<CodeEditor>;
+export type MToolbarListButtonData = ToolbarListButtonData<CodeEditor>;
+export type MToolbarListItemData = ToolbarListItemData<CodeEditor>;
+export type MToolbarButtonPopupData = ToolbarButtonPopupData<CodeEditor>;
