+++ src/bundle/toolbar/utils/flattenPreset.ts
@@ -1,6 +1,5 @@
-import {ToolbarDataType} from '../../../toolbar/types';
-import type {MToolbarData, MToolbarItemData} from '../../config';
-import type {WToolbarData, WToolbarItemData} from '../types';
+import type {MToolbarData, MToolbarItemData, WToolbarData, WToolbarItemData} from '../types';
+import {ToolbarDataType} from '../types';
 
 export const flattenPreset = <T extends WToolbarData | MToolbarData>(
     config: T,
