+++ src/bundle/toolbar/wysiwyg/WToolbarTextSelect.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import {ActionStorage} from '../../../core';
-import {ToolbarBaseProps} from '../../../toolbar/types';
+import type {ActionStorage} from '../../../core';
+import type {ToolbarBaseProps} from '../../../toolbar/types';
 import {wHeadingListConfig} from '../../config/w-heading-config';
 import {ToolbarSelect} from '../ToolbarSelect';
