+++ src/extensions/additional/GPT/MarkupGpt/toolbar.ts
@@ -1,7 +1,6 @@
-import {MToolbarSingleItemData} from '../../../../bundle/config/markup';
+import {type MToolbarSingleItemData, ToolbarDataType} from '../../../../bundle/toolbar/types';
 import {i18n} from '../../../../i18n/gpt/extension';
 import {GPTIcon} from '../../../../icons';
-import {ToolbarDataType} from '../../../../toolbar';
 import {gptHotKeys} from '../constants';
 
 import {showMarkupGpt} from './commands';
