+++ src/extensions/yfm/YfmTable/YfmTableSpecs/index.ts
@@ -1,23 +1,17 @@
 import log from '@diplodoc/transform/lib/log';
 import yfmTable from '@diplodoc/transform/lib/plugins/table';
-import type {NodeSpec} from 'prosemirror-model';
 
 import type {ExtensionWithOptions} from '../../../../core';
 
 import {YfmTableNode} from './const';
 import {parserTokens} from './parser';
-import {getSchemaSpecs} from './schema';
+import {type YfmTableSchemaOptions, getSchemaSpecs} from './schema';
 import {serializerTokens} from './serializer';
 
 export {YfmTableNode} from './const';
 export {yfmTableType, yfmTableBodyType, yfmTableRowType, yfmTableCellType} from './utils';
 
-export type YfmTableSpecsOptions = {
-    /**
-     * @deprecated: use placeholder option in BehaviorPreset instead.
-     */
-    yfmTableCellPlaceholder?: NonNullable<NodeSpec['placeholder']>['content'];
-};
+export type YfmTableSpecsOptions = YfmTableSchemaOptions & {};
 
 export const YfmTableSpecs: ExtensionWithOptions<YfmTableSpecsOptions> = (builder, options) => {
     const schemaSpecs = getSchemaSpecs(options, builder.context.get('placeholder'));
