+++ src/markup/codemirror/html-to-markdown/handlers.ts
@@ -1,4 +1,4 @@
-import {HTMLNodeVisitor} from './converters';
+import type {HTMLNodeVisitor} from './converters';
 
 /**
  * Base handler class implementing the Chain of Responsibility pattern for HTML node processing.
