+++ src/react-utils/react-node-view.tsx
@@ -4,8 +4,8 @@ import type {Mark, Node} from 'prosemirror-model';
 import type {EditorView, NodeView, NodeViewConstructor} from 'prosemirror-view';
 import {createPortal} from 'react-dom';
 
-import {ExtensionDeps, Serializer} from '../core';
-import {getReactRendererFromState} from '../extensions';
+import type {ExtensionDeps, Serializer} from '../core';
+import {getReactRendererFromState} from '../extensions/behavior/ReactRenderer';
 
 type ReactNodeViewOptions<T> = {
     isInline?: boolean;
